+++ src/lib/proof-system/proof.ts
@@ -2,7 +2,7 @@ import { initializeBindings, withThreadPool } from '../../snarky.js';
 import { Pickles } from '../../snarky.js';
 import { Field, Bool } from '../provable/wrapped.js';
 import type {
-  FlexibleProvablePure,
+  FlexibleProvable,
   InferProvable,
 } from '../provable/types/struct.js';
 import { FeatureFlags } from './feature-flags.js';
@@ -25,8 +25,8 @@ type MaxProofs = 0 | 1 | 2;
 type ProofClass = Subclass<typeof ProofBase>;
 
 class ProofBase<Input = any, Output = any> {
-  static publicInputType: FlexibleProvablePure<any> = undefined as any;
-  static publicOutputType: FlexibleProvablePure<any> = undefined as any;
+  static publicInputType: FlexibleProvable<any> = undefined as any;
+  static publicOutputType: FlexibleProvable<any> = undefined as any;
   static tag: () => { name: string } = () => {
     throw Error(
       `You cannot use the \`Proof\` class directly. Instead, define a subclass:\n` +